add chart continuation and custom chart save/open path in cfg file
also removes press 's' to save since it doesn't apparently work and delays the program by 20ms each time it runs; also removes the -chart option. to specify a chart path when training, add an option called "chart" in the data file, which specifies the chart path. if no chart path is specified, chart.png is saved in the current directory.